Update raster style

+++ app/layers/views.py
@@ -8,6 +8,7 @@ from app.layers.models import (
 from app.db import get_session, AsyncSession
 from fastapi import (
     Depends,
+    HTTPException,
     APIRouter,
     Query,
     BackgroundTasks,
@@ -26,6 +27,8 @@ from app.layers.services import (
 )
 from typing import Any
 from sqlmodel import select
+import httpx
+from app.config import config
 
 router = APIRouter()
 
@@ -41,6 +44,7 @@ async def get_groups(
     groups = {}
     for group in LayerVariables:
         # Get distinct values for each group
+
         column = getattr(Layer, group.value)
         res = await session.exec(select(column).distinct())
         groups[group.value] = [row for row in res.all()]
@@ -90,31 +94,48 @@ async def update_layer(
     return updated_layer
 
 
-@router.delete("/batch", response_model=list[str])
-async def delete_batch(
-    ids: list[UUID],
-    session: AsyncSession = Depends(get_session),
-) -> list[str]:
-    """Delete by a list of ids"""
+# @router.delete("/batch", response_model=list[str])
+# async def delete_batch(
+#     ids: list[UUID],
+#     session: AsyncSession = Depends(get_session),
+# ) -> list[str]:
+#     """Delete by a list of ids"""
 
-    for id in ids:
-        obj = await crud.get_model_by_id(model_id=id, session=session)
-        if obj:
-            await session.delete(obj)
+#     for id in ids:
+#         obj = await crud.get_model_by_id(model_id=id, session=session)
+#         if obj:
+#             await session.delete(obj)
 
-    await session.commit()
+#     await session.commit()
 
-    return [str(obj_id) for obj_id in ids]
+#     return [str(obj_id) for obj_id in ids]
 
 
 @router.delete("/{layer_id}")
 async def delete_layer(
     layer: LayerRead = Depends(get_one),
     session: AsyncSession = Depends(get_session),
-) -> None:
+) -> str:
     """Delete a layer by id"""
+
+    id = layer.id
+    # First delete the coveragestore from geoserver
+    # Then delete the layer from the database
+    async with httpx.AsyncClient() as client:
+        url = (
+            f"{config.GEOSERVER_URL}/rest/workspaces/"
+            f"{config.GEOSERVER_WORKSPACE}/coveragestores/{layer.layer_name}"
+        )
+        auth = (config.GEOSERVER_USER, config.GEOSERVER_PASSWORD)
+        res = await client.delete(url, auth=auth, params={"recurse": "true"})
+
+    if res.status_code > 299:
+        raise HTTPException(
+            status_code=res.status_code,
+            detail=res.text,
+        )
 
     await session.delete(layer)
     await session.commit()
 
-    return {"ok": True}
+    return id
